+++ src/RcsbFv/RcsbFvConfig/RcsbFvDOMConstants.ts
@@ -2,6 +2,7 @@
 export const enum RcsbFvDOMConstants {
     UI_DOM_ID_PREFIX = "_uiDiv",
     GLOW_DOM_ID_PREFIX = "_glowDiv",
+    GLOW_ROW_DOM_ID_SUFFIX = "_glowRowDiv",
     TOOLTIP_DOM_ID_PREFIX = "_tooltip",
     TOOLTIP_DESCRIPTION_DOM_ID_PREFIX = "_tooltipDescription",
     PROGRESS_DIV_DOM_ID_PREFIX = "_progressDiv",